Combine equalizer filters to a single command
